added support for IF NOT EXISTS
From: Felipe Sateler <fsateler@debian.org>
Subject: Bug#671395: apgdiff: Support IF NOT EXISTS in CREATE TABLE

apgdiff chokes if it encounters a CREATE TABLE IF NOT EXISTS statement.
The attached patch makes it discard the IF NOT EXISTS section if it is
found. The IF NOT EXISTS portion is irrelevant for apgdiff, so we do
nothing with it.